user delete fix
default user default gets deleted after user delete

+++ backend/controllers/userController.js
@@ -331,7 +331,7 @@ const deleteUser = asyncHandler(async (req, res) => {
   }
 
   const destination = "frontend\\public";
-  if (user.image) {
+  if (user.image && user.image !== "\\uploads\\users\\user-placeholder.png") {
     fs.unlinkSync(destination + user.image);
   }
